envoy: Bump to 1.22.6
tcmalloc needed an update to compile on arm64.

abseil needed a patch to avoid a warnign on use of an uninitialized
variable.

boringssl BUILD file needed a patch to be compatible with old-style
cpu-flag for cross-compilation.

tests needed cert updates and change from std::string.substr() to
string_view.substr(), this fixed one test case seeing garbage in the
data.

+++ cilium/websocket_codec.cc
@@ -204,7 +204,7 @@ public:
 
   unsigned int status() {
     ENVOY_LOG(trace, "websocket: http_parser got status: {}",
-	      parser_.status_code);
+	      static_cast<unsigned int>(parser_.status_code));
     return parser_.status_code;
   }
